StoryTellerSDK/src/main/server/requirements.txt
alembic==1.4.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
conda==4.8.3
conda-package-handling==1.7.0+0.g7c4a471.dirty
cryptography==2.9.2
cycler==0.10.0
databricks-cli==0.14.1
decorator==4.4.2
dill==0.3.3
docker==5.0.0
entrypoints==0.3
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
gunicorn==20.1.0
idna==2.9
importlib-metadata==4.6.4
iniconfig==1.1.1
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
kiwisolver==1.3.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.2
mlflow==1.20.1
networkx==2.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
parso==0.8.1
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
py==1.10.0
py4j==0.10.9
pycosat==0.6.3
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
querystring-parser==1.2.4
requests==2.23.0
ruamel-yaml==0.15.87
scipy==1.6.0
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.2
torch==1.10.0
tqdm==4.46.0
typing-extensions==3.7.4.3
urllib3==1.25.8
websocket-client==1.2.1
Werkzeug==2.0.1
zipp==3.5.0
